move transformation tabs out of folder
